Update streamlit_app.py
In this step, we will move several lines of code into a little group of code called a function. Functions can be used to make your code more organized. A function begins with the word "def" and is followed by the name of the function -- which you decide. In this case, you can see on line 29 of the AFTER section, it says def get_fruityvice_data -- that means "define what get_fruityvice_data is going to mean".